Remove KaitaiStruct.ReadWrite._toByteArray() method
It's not clear how exactly it should behave (should it write only `seq`
fields, or `instances` as well?) and how to correctly implement it (we
don't even have any concept of a "growable" stream yet). The type we
would be writing might also depend on some `_io` properties (of its own
`_io`, or of an other object's `_io`), which are beyond our capabilities
to set like they would be when writing the type for real (for starters,
we don't know at which `_io.pos` the type starts in the actual format).

So let's not pretend that things are easier than they are. If users
need this functionality, they should do it themselves.