Update pos() and isEof() for {read,write}BitsInt*() parity
If you are at `pos() == 0` on a byte boundary and call readBitsInt*(2),
bb.position() is now 1. However, calling writeBitsInt*(2) only modifies
the internal `bitsLeft`, `bits` fields and does not shift bb.position()
just yet, so it stays at `bb.position() == 0`.

isEof() has to be changed too. If there are no bytes remaining
(`!bb.hasRemaining()`) after a readBitsInt*() call, it only tells that
the last byte has been _read_, but we're only at EOF when `bitsLeft ==
0`, i.e. when the bit buffer becomes empty.

However, no bytes remaining after a writeBitsInt*() call means that the
last available byte has been fully _written_, so the stream can't hold
even a single extra bit. But currently there are no checks in
writeBitsInt*() on whether it will be possible to write the bits it
saves for later, so if `!bb.hasRemaining() && bitsLeft > 0` after a
writeBitsInt*() call, we're actually _past_ EOF and isEof() should
return true just because of `!bb.hasRemaining()`.

The `!bb.hasRemaining() && bitsLeft > 0` state after writeBitsInt*()
creates a weird situation where `pos() == size() + 1`, though. But
that's at least in a sense correct - the stream shouldn't have been
allowed to get there in the first place, that's the confusing thing, but
I don't think there's anything to change about the reported values.