RAF KaitaiStream: fix premature throwing of EOFException
RandomAccessFile.read() can read fewer bytes than requested, it does not
necessarily mean that we are close to EOF. Therefore, our existing
implementation of readBytesNotAligned() could throw an EOFException
prematurely.

To fix this, we use
[RandomAccessFile.readFully()](https://docs.oracle.com/javase/7/docs/api/java/io/RandomAccessFile.html#readFully%28byte%5B%5D%29),
which internally repeatedly calls RandomAccessFile.read() until the
required number of bytes is read, or throws an EOFException itself if
EOF has actually been reached.